requirements.txt
# From Docker documentation:
# https://docs.docker.com/samples/django/

asgiref>=3.4.1
Django>=3.2.9
Pillow>=3.0.0
pip>=21.2.4
psycopg2>=2.9.2
pycryptodome>=3.12.0
pytz>=2021.3
setuptools>=57.5.0
sqlparse>=0.4.2
wheel>=0.24.0